Sigo indicaciones revisor. Ayuda a resolver #7, #8 y #9
