+++ pyproyect.toml
@@ -1,0 +1,14 @@
+[build-system]
+requires = ["setuptools", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "Horarios Automatrícula"  # Nombre del proyecto
+version = "0.1.0"  # Versión del proyecto
+description = "Sistema para gestionar horarios en la automatrícula de estudiantes."
+readme = "README.md"  # Archivo README que describirá el proyecto
+license = { text = "MIT" }  # Tipo de licencia
+requires-python = ">=3.8"  # Versión mínima de Python requerida
+
+[tool.setuptools]
+packages = ["horarios_automatricula"]  # Definimos el paquete en la carpeta horarios_automatricula
